Move decision alert to files page (#4785)
* Move decision alert to files page

* Go to files tab for spec

* Lint fix

* Accessibility spec update

* Accessibility spec fix
