+++ test/accessibility/disability-benefits-508.spec.js
@@ -43,6 +43,12 @@ module.exports = E2eHelpers.createE2eTest(
       .click('.va-nav-breadcrumbs-list li:nth-child(2) a')
       .waitForElementVisible('.claim-title', Timeouts.normal);
 
+    // files tab
+    client
+      .click('.va-tabs li:nth-child(2) > a')
+      .waitForElementVisible('.file-request-list-item', Timeouts.normal)
+      .axeCheck('.main');
+
     // Ask VA page
     client
       .click('.usa-alert-info a')
@@ -51,13 +57,9 @@ module.exports = E2eHelpers.createE2eTest(
 
     client
       .click('.va-nav-breadcrumbs-list li:nth-child(2) a')
-      .waitForElementVisible('.claim-title', Timeouts.normal);
-
-    // files tab
-    client
+      .waitForElementVisible('.claim-title', Timeouts.normal)
       .click('.va-tabs li:nth-child(2) > a')
-      .waitForElementVisible('.file-request-list-item', Timeouts.normal)
-      .axeCheck('.main');
+      .waitForElementVisible('.file-request-list-item', Timeouts.normal);
 
     // document request page
     client
