+++ test/disability-benefits/components/StatusPage.unit.spec.jsx
@@ -28,7 +28,6 @@ describe('<StatusPage>', () => {
           claim={claim}/>
     );
     expect(tree.subTree('NeedFilesFromYou')).to.be.false;
-    expect(tree.subTree('AskVAToDecide')).to.be.false;
     expect(tree.subTree('ClaimsDecision')).to.be.false;
     expect(tree.subTree('ClaimsTimeline')).not.to.be.false;
   });
@@ -150,30 +149,6 @@ describe('<StatusPage>', () => {
           claim={claim}/>
     );
     expect(tree.subTree('NeedFilesFromYou')).to.be.false;
-  });
-  it('should render ask va to decide component', () => {
-    const claim = {
-      id: 1,
-      attributes: {
-        phase: 3,
-        documentsNeeded: false,
-        decisionLetterSent: false,
-        waiverSubmitted: false,
-        eventsTimeline: [
-          {
-            type: 'still_need_from_you_list',
-            status: 'NEEDED'
-          }
-        ]
-      }
-    };
-
-    const tree = SkinDeep.shallowRender(
-      <StatusPage
-          params={{ id: 2 }}
-          claim={claim}/>
-    );
-    expect(tree.everySubTree('AskVAToDecide')).not.to.be.empty;
   });
   it('should render claims decision alert', () => {
     const claim = {
